Refactor screens home
