+++ app/src/main/java/com/afoxplus/yalistoadmin/ui/screens/home/YaListoHomeState.kt
@@ -1,4 +1,4 @@
-package com.afoxplus.yalistoadmin.ui.main
+package com.afoxplus.yalistoadmin.ui.screens.home
 
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
@@ -13,7 +13,6 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.afoxplus.yalistoadmin.commons.extensions.navigateSingleTopTo
-import com.afoxplus.yalistoadmin.ui.screens.home.BottomBarHomeRouter
 import kotlinx.coroutines.CoroutineScope
 
 @Composable
