+++ app/src/main/java/com/afoxplus/yalistoadmin/ui/screens/order_status/components/OrderShareView.kt
@@ -1,4 +1,4 @@
-package com.afoxplus.yalistoadmin.ui.screens.details.components
+package com.afoxplus.yalistoadmin.ui.screens.order_status.components
 
 import android.content.Context
 import android.content.Intent
