+++ app/src/main/java/com/afoxplus/yalistoadmin/ui/screens/order_status/components/OrderWhatsappContactComponent.kt
@@ -1,4 +1,4 @@
-package com.afoxplus.yalistoadmin.ui.screens.details.components
+package com.afoxplus.yalistoadmin.ui.screens.order_status.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
