+++ app/src/main/java/com/afoxplus/yalistoadmin/ui/screens/orders/components/OrdersComponent.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.afoxplus.uikitcompose.ui.theme.Light01
 import com.afoxplus.yalistoadmin.domain.entities.Order
@@ -36,10 +35,4 @@ fun OrdersComponent(
             )
         }
     }
-}
-
-@Preview(showSystemUi = true)
-@Composable
-fun OrdersComponentPreview() {
-    OrdersComponent(orders = arrayListOf()) {}
 }
